requirements.txt
fastapi
uvicorn
pydantic
pytest
httpx
sentence-transformers
faiss-cpu